# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentacion",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    }
  ],
  "paths": {
...
    "/usuarios/crear": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Crear usuarios",
        "operationId": "Crearusuarios",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CrearusuariosRequest"
                  },
                  {
                    "example": {
                      "usuario": "alex10",
                      "clave": "alexander10",
                      "nombre": "Alexander"
                    }
                  }
                ]
              },
              "example": {
                "usuario": "alex10",
                "clave": "alexander10",
                "nombre": "Alexander"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/usuarios/login": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Login",
        "operationId": "Login",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LoginRequest"
                  },
                  {
                    "example": {
                      "usuario": "alex10",
                      "clave": "alexander10"
                    }
                  }
                ]
              },
              "example": {
                "usuario": "alex10",
                "clave": "alexander10"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/pacientes/crear": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Crear pacientes",
        "operationId": "Crearpacientes",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CrearpacientesRequest"
                  },
                  {
                    "example": {
                      "apellidos": "Delgado Tapia",
                      "nombres": "José Roberto",
                      "edad": "55",
                      "sexo": "Masculino",
                      "odontograma": "000"
                    }
                  }
                ]
              },
              "example": {
                "apellidos": "Delgado Tapia",
                "nombres": "José Roberto",
                "edad": "55",
                "sexo": "Masculino",
                "odontograma": "000"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/tratamientos/crear": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Crear tratamientos",
        "operationId": "Creartratamientos",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreartratamientosRequest"
                  },
                  {
                    "example": {
                      "fecha": "2024-07-29T05:00:00.000+00:00",
                      "tratamiento": "Restauracion",
                      "diagnostico": "Caries en terce molar izquierdo",
                      "idpaciente": 8,
                      "idtipo": 3
                    }
                  }
                ]
              },
              "example": {
                "fecha": "2024-07-29T05:00:00.000+00:00",
                "tratamiento": "Restauracion",
                "diagnostico": "Caries en terce molar izquierdo",
                "idpaciente": 8,
                "idtipo": 3
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "CrearusuariosRequest": {
        "title": "CrearusuariosRequest",
        "required": [
          "usuario",
          "clave",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "usuario": {
            "type": "string"
          },
          "clave": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          }
        },
        "example": {
          "usuario": "alex10",
          "clave": "alexander10",
          "nombre": "Alexander"
        }
      },
      "LoginRequest": {
        "title": "LoginRequest",
        "required": [
          "usuario",
          "clave"
        ],
        "type": "object",
        "properties": {
          "usuario": {
            "type": "string"
          },
          "clave": {
            "type": "string"
          }
        },
        "example": {
          "usuario": "alex10",
          "clave": "alexander10"
        }
      },
      "CrearpacientesRequest": {
        "title": "CrearpacientesRequest",
        "required": [
          "apellidos",
          "nombres",
          "edad",
          "sexo",
          "odontograma"
        ],
        "type": "object",
        "properties": {
          "apellidos": {
            "type": "string"
          },
          "nombres": {
            "type": "string"
          },
          "edad": {
            "type": "string"
          },
          "sexo": {
            "type": "string"
          },
          "odontograma": {
            "type": "string"
          }
        },
        "example": {
          "apellidos": "Delgado Tapia",
          "nombres": "José Roberto",
          "edad": "55",
          "sexo": "Masculino",
          "odontograma": "000"
        }
      },
...
      "CreartratamientosRequest": {
        "title": "CreartratamientosRequest",
        "required": [
          "fecha",
          "tratamiento",
          "diagnostico",
          "idpaciente",
          "idtipo"
        ],
        "type": "object",
        "properties": {
          "fecha": {
            "type": "string"
          },
          "tratamiento": {
            "type": "string"
          },
          "diagnostico": {
            "type": "string"
          },
          "idpaciente": {
            "type": "integer",
            "format": "int32"
          },
          "idtipo": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "fecha": "2024-07-29T05:00:00.000+00:00",
          "tratamiento": "Restauracion",
          "diagnostico": "Caries en terce molar izquierdo",
          "idpaciente": 8,
          "idtipo": 3
        }
      }
    }
  },
...
}